MANY IMPROVEMENTS AND TEXTcd ..! Optimization with Meshes, need to optimize Mesh gen

+++ src/utils.cpp
@@ -1,10 +1,10 @@
 #include "utils.h"
 
 void split(std::string* bufstr, std::string str, char separator) {
-    uint16_t len = strlen(str.data());
-    uint8_t d = 0;
+    uint len = strlen(str.data());
+    uint d = 0;
     std::string s;
-    for (uint8_t i = 0; i < len; i++) {
+    for (uint i = 0; i < len; i++) {
         if (str.data()[i] != separator) s += str.data()[i];
         else {
             bufstr[d] = s;
